+++ middleware_api/lambda/inferences/inference_api.py
@@ -14,9 +14,9 @@ from sagemaker.serializers import JSONSerializer
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request
 from common.util import generate_presign_url, load_json_from_s3, upload_json_to_s3, split_s3_path
-from common.types import CheckPoint, CheckPointStatus
-from common.types import InferenceJob, InvocationsRequest, EndpointDeploymentJob
-from common.utils import get_user_roles, check_user_permissions
+from libs.data_types import CheckPoint, CheckPointStatus
+from libs.data_types import InferenceJob, InvocationsRequest, EndpointDeploymentJob
+from libs.utils import get_user_roles, check_user_permissions
 
 bucket_name = os.environ.get('S3_BUCKET')
 checkpoint_table = os.environ.get('CHECKPOINT_TABLE')
