+++ middleware_api/lambda/libs/utils.py
@@ -3,7 +3,7 @@ import logging
 import boto3
 from botocore.exceptions import ClientError
 
-from types import PARTITION_KEYS, User, Role
+from libs.data_types import PARTITION_KEYS, User, Role
 
 
 class KeyEncryptService:
