+++ middleware_api/lambda/models/create_model.py
@@ -1,5 +1,4 @@
 import dataclasses
-import dataclasses
 import datetime
 import json
 import logging
@@ -8,12 +7,12 @@ from typing import Any, Optional
 
 from botocore.exceptions import ClientError
 
-from common.common_tools import get_base_model_s3_key, get_base_checkpoint_s3_key, \
+from libs.common_tools import get_base_model_s3_key, get_base_checkpoint_s3_key, \
     batch_get_s3_multipart_signed_urls
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request, internal_server_error
-from common.types import Model, CreateModelStatus, CheckPoint, CheckPointStatus, MultipartFileReq
-from common.utils import get_permissions_by_username, get_user_roles
+from libs.data_types import Model, CreateModelStatus, CheckPoint, CheckPointStatus, MultipartFileReq
+from libs.utils import get_permissions_by_username, get_user_roles
 
 bucket_name = os.environ.get('S3_BUCKET')
 model_table = os.environ.get('DYNAMODB_TABLE')
