+++ middleware_api/lambda/trainings/create_training_job.py
@@ -10,10 +10,10 @@ import boto3
 
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request, not_found, forbidden, internal_server_error
-from common.types import TrainJob, TrainJobStatus, Model, CreateModelStatus, CheckPoint, CheckPointStatus
+from libs.data_types import TrainJob, TrainJobStatus, Model, CreateModelStatus, CheckPoint, CheckPointStatus
 from common.util import get_s3_presign_urls
 from common.util import load_json_from_s3, save_json_to_file
-from common.utils import get_permissions_by_username, get_user_roles
+from libs.utils import get_permissions_by_username, get_user_roles
 
 bucket_name = os.environ.get('S3_BUCKET')
 train_table = os.environ.get('TRAIN_TABLE')
